refactor: token payload 형식 변경
