+++ src/auth/jwt-generate.service.ts
@@ -1,3 +1,4 @@
+import { TokenPayload } from '#auth/auth.types.js';
 import { IJwtGenerateService } from '#auth/interfaces/jwt-generate.service.interface.js';
 import { UserRepository } from '#users/user.repository.js';
 import { Injectable } from '@nestjs/common';
@@ -12,7 +13,7 @@ export class JwtGenerateService implements IJwtGenerateService {
     private readonly configService: ConfigService,
   ) {}
 
-  generateAccessToken(payload: { userId: string }) {
+  generateAccessToken(payload: TokenPayload) {
     const jwtSecret = this.configService.get('jwtSecret');
     const accessExpireTime = this.configService.get('accessExpireTime');
 
@@ -23,7 +24,7 @@ export class JwtGenerateService implements IJwtGenerateService {
     return accessToken;
   }
 
-  async generateRefreshToken(payload: { userId: string }) {
+  async generateRefreshToken(payload: TokenPayload) {
     const jwtSecret = this.configService.get('jwtSecret');
     const refreshExpireTime = this.configService.get('refreshExpireTime');
 
@@ -31,7 +32,7 @@ export class JwtGenerateService implements IJwtGenerateService {
       secret: jwtSecret,
       expiresIn: refreshExpireTime,
     });
-    await this.userRepository.update(payload.userId, { refreshToken });
+    await this.userRepository.update(payload.id, { refreshToken });
     return refreshToken;
   }
 }
